+++ assets/javascripts/kitten/components/form/form-amount-and-currency.test.js
@@ -6,7 +6,7 @@ import { TextInput } from 'kitten/components/form/text-input'
 import { FormAmountAndCurrency } from 'kitten/components/form/form-amount-and-currency'
 
 describe('<FormAmountAndCurrency />', () => {
-  describe('valid props', () => {
+  describe('valid prop', () => {
     const component = mount(<FormAmountAndCurrency valid={ true } />)
     const textInput = component.find('input')
     const textInputExpectation = textInput.hasClass('is-valid')
@@ -16,7 +16,7 @@ describe('<FormAmountAndCurrency />', () => {
     })
   })
 
-  describe('error props', () => {
+  describe('error prop', () => {
     const component = mount(<FormAmountAndCurrency error={ true } />)
     const textInput = component.find('input')
     const textInputExpectation = textInput.hasClass('is-error')
@@ -26,7 +26,7 @@ describe('<FormAmountAndCurrency />', () => {
     })
   })
 
-  describe('tiny props', () => {
+  describe('tiny prop', () => {
     const component = mount(<FormAmountAndCurrency tiny={ true } />)
     const textInput = component.find('input')
     const select = component.find('.k-Select')
@@ -42,20 +42,17 @@ describe('<FormAmountAndCurrency />', () => {
     })
   })
 
-  describe('disabled props', () => {
-    const component = mount(<FormAmountAndCurrency disabled={ true } />)
-    const textInput = component.find('input')
+  describe('disabled prop', () => {
+    const component = shallow(<FormAmountAndCurrency disabled={ true } />)
+    const textInput = component.find('TextInput')
+    const select = component.find('SelectWithState')
 
-    const expectedProps = {
-      disabled: true,
-    }
-
-    it('renders a disabled <input />', () => {
-      expect(textInput).to.have.attr('disabled')
+    it('passes the right props to the `TextInput` component', () => {
+      expect(textInput.props()).to.contains.all.keys({ disabled: true })
     })
 
-    it('passes the right props to the `Select` component', () => {
-      expect(component.props()).to.contain.all.keys(expectedProps)
+    it('passes the right props to the `SelectWithState` component', () => {
+      expect(select.props()).to.contains.all.keys({ disabled: true })
     })
   })
 })