+++ CHANGELOG.md
@@ -4,9 +4,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Breaking change: the component `FormAmountAndCurrency` is not tiny by default.
-- Breaking change: the component `SelectWithState` has a `tiny` props instead
-  of a `big` one.
+- Breaking change: The `FormAmountAndCurrency` has new props.
+  To be retro-compatible you should, in your app's components:
+  - Rename `value` to `currencyValue`.
+  - Rename `amount` to `currencyAmount`.
+  - Add the `tiny` boolean prop.
+- Breaking change: The `SelectWithState` has a new default size and export.
+  To be retro-compatible you should, in your app's components:
+  - Add the `tiny` boolean prop if the component does not have `big` set
+    to `true`.
+  - Remove the `big` boolean prop.
+  - Replace `import FormAmountAndCurrency` by
+    `import { FormAmountAndCurrency }`.
 
 ## [6.4.0] - 2017-02-20